requirements.txt
alembic==1.17.2
annotated-types==0.7.0
anyio==4.12.0
bidict==0.23.1
black==25.12.0
certifi==2025.11.12
click==8.3.1
granian==2.6.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.1
python-engineio==4.13.0
python-multipart==0.0.21
python-socketio==5.16.0
pytokens==0.3.0
redis==7.1.0
reflex==0.8.24.post1
reflex-hosting-cli==0.1.61
rich==14.2.0
simple-websocket==1.1.0
SQLAlchemy==2.0.45
sqlmodel==0.0.31
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
watchfiles==1.1.1
wrapt==2.0.1
wsproto==1.3.2
